requirements.txt
crewai==0.70.1
python-dotenv==1.0.0
groq==0.9.0
streamlit==1.28.0